garmindb/pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "FitnessTracker"
version = "0.1.0"
description = "Flask application for GarminDB"
authors = [
    { name = "Samuli Järvensivu", email = "samuli.jarvensivu@gmail.com" }
]
dependencies = [
    "SQLAlchemy==2.0.27",
    "python-dateutil==2.8.2",
    "cached-property==1.5.2",
    "tqdm==4.66.2",
    "garth>=0.4.44",
    "fitfile>=1.1.7",
    "tcxfile>=1.0.4",
    "idbutils>=1.1.0",
    "Flask==3.0.3",
    "Flask-Cors==4.0.1",
    "Flask-PyMongo==2.3.0",
    "Flask-SocketIO==5.3.6",
    "Flask-SQLAlchemy==3.1.1",
    "garminconnect==0.2.19",
    "garmindb==3.6.0",
    "pyodbc==5.1.0",
    "python-dotenv==1.0.1",
    "setuptools==71.1.0",
    "webencodings==0.5.1",
    "websocket-client==1.8.0"
]

[tool.setuptools]
packages = ["FitnessTracker"]

[tool.setuptools.package-data]
"FitnessTracker" = ["*.txt", "*.rst", "*.md"]
